nixos/test-driver: make the test-driver a python package
